+++ R/onLoad.R
@@ -4,7 +4,6 @@
 #'
 #' @noRd
 .onLoad <- function(...) {
-  shiny::addResourcePath("shinylogs", system.file("assets", package = "shinylogs"))
-  shiny::registerInputHandler("parse_logInput", parse_logInput)
-  shiny::registerInputHandler("parse_lastInput", parse_lastInput)
+  shiny::registerInputHandler("parse_logInput", parse_logInput, force = TRUE)
+  shiny::registerInputHandler("parse_lastInput", parse_lastInput, force = TRUE)
 }